+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_execute_schedule.py
@@ -6,6 +6,7 @@ from dagster_graphql.test.utils import define_context_for_repository_yaml, execu
 
 from dagster import seven
 from dagster.core.instance import DagsterInstance, InstanceType
+from dagster.core.scheduler import reconcile_scheduler_state
 from dagster.core.scheduler.scheduler import ScheduleTickStatus
 from dagster.core.storage.event_log import InMemoryEventLogStorage
 from dagster.core.storage.local_compute_log_manager import NoOpComputeLogManager
@@ -348,11 +349,8 @@ def test_tick_success(snapshot):
         )
         repository = context.get_repository()
 
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
-        schedule_def = schedule_handle.get_schedule_def_by_name(
-            "no_config_pipeline_hourly_schedule"
-        )
+        reconcile_scheduler_state("", "", repository, instance)
+        schedule_def = repository.get_schedule_def("no_config_pipeline_hourly_schedule")
 
         start_time = time.time()
         execute_dagster_graphql(
@@ -389,8 +387,7 @@ def test_tick_skip(snapshot):
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         execute_dagster_graphql(
             context,
@@ -422,8 +419,7 @@ def test_should_execute_scheduler_error(snapshot):
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         execute_dagster_graphql(
             context,
@@ -460,8 +456,7 @@ def test_tags_scheduler_error(snapshot):
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         result = execute_dagster_graphql(
             context,
@@ -497,8 +492,7 @@ def test_enviornment_dict_scheduler_error(snapshot):
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         result = execute_dagster_graphql(
             context,
@@ -535,8 +529,7 @@ def test_enviornment_dict_scheduler_error_serialize_cauze():
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         result = execute_dagster_graphql(
             context,
@@ -563,8 +556,7 @@ def test_query_multiple_schedule_ticks(snapshot):
             path=file_relative_path(__file__, '../repository.yaml'), instance=instance
         )
         repository = context.get_repository()
-        schedule_handle = context.scheduler_handle
-        schedule_handle.up("", "", repository, instance)
+        reconcile_scheduler_state("", "", repository, instance)
 
         for scheduleName in [
             'no_config_pipeline_hourly_schedule',