+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_scheduler.py
@@ -6,7 +6,12 @@ from dagster_graphql.test.utils import define_context_for_repository_yaml, execu
 
 from dagster import ScheduleDefinition, seven
 from dagster.core.instance import DagsterInstance, InstanceType
-from dagster.core.scheduler import Schedule, ScheduleStatus, get_schedule_change_set
+from dagster.core.scheduler import (
+    Schedule,
+    ScheduleStatus,
+    get_schedule_change_set,
+    reconcile_scheduler_state,
+)
 from dagster.core.storage.event_log import InMemoryEventLogStorage
 from dagster.core.storage.local_compute_log_manager import NoOpComputeLogManager
 from dagster.core.storage.root import LocalArtifactStorage
@@ -69,8 +74,7 @@ def test_get_all_schedules():
 
         # Initialize scheduler
         repository = context.get_repository()
-        scheduler_handle = context.scheduler_handle
-        scheduler_handle.up(
+        reconcile_scheduler_state(
             python_path=sys.executable,
             repository_path="",
             repository=repository,